notebook/C7 高级 RAG 技巧/7. 评估/requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.35.25
botocore==1.35.25
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cohere==5.9.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
courlan==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
Faker==25.9.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.7
favicon==0.7.0
filelock==3.15.1
FlagEmbedding==1.2.10
flatbuffers==24.3.25
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
htbuilder==0.6.2
html2text==2024.2.26
htmldate==1.8.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.3
humanfriendly==10.0
humanize==4.9.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
jusText==3.0.1
kaggle==1.6.17
kiwisolver==1.4.5
kubernetes==30.1.0
langchain
langchain-community
langchain-core==0.1.52
langchain-openai
langchain-text-splitters
langsmith==0.1.77
litellm==1.48.0
llama-index==0.10.48.post1
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.48
llama-index-embeddings-huggingface==0.2.1
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-anyscale==0.1.4
llama-index-llms-fireworks==0.1.5
llama-index-llms-langchain==0.1.4
llama-index-llms-openai==0.1.22
llama-index-llms-replicate==0.1.3
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-chroma==0.1.9
llama-index-vector-stores-faiss==0.1.2
llama-parse==0.4.4
llamaindex-py-client==0.1.19
lxml==5.2.2
lxml_html_clean==0.1.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
minijinja==2.0.1
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.47.1
openpyxl==3.1.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.4
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pillow==10.3.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
posthog==3.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pymdown-extensions==10.8.1
Pympler==1.0.1
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
semver==3.0.2
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.16.0
slack_bolt==1.20.1
slack_sdk==3.33.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
streamlit-aggrid==0.3.4
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-javascript==0.1.5
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
striprtf==0.0.26
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tld==0.13
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
trafilatura==1.10.0
traitlets==5.14.3
transformers==4.41.2
trulens_eval==0.31.0
typer==0.12.3
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
validators==0.28.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zhipuai==2.1.1.20240620.1
zipp==3.19.2
